+++ app/src/main/java/com/cmoney/backend2/sample/servicecase/ProfileServiceCase.kt
@@ -1,7 +1,10 @@
 package com.cmoney.backend2.sample.servicecase
 
+import android.util.Log
 import com.cmoney.backend2.profile.service.ProfileWeb
 import com.cmoney.backend2.profile.service.api.getusergraphqlinfo.UserGraphQLInfo
+import com.cmoney.backend2.profile.service.api.mutationmyusergraphqlinfo.City
+import com.cmoney.backend2.profile.service.api.mutationmyusergraphqlinfo.Education
 import com.cmoney.backend2.profile.service.api.mutationmyusergraphqlinfo.MutationData
 import com.cmoney.backend2.profile.service.api.variable.GraphQLFieldDefinition
 import com.cmoney.backend2.sample.extension.logResponse
@@ -53,16 +56,33 @@ class ProfileServiceCase : ServiceCase {
         ).logResponse(TAG)
 
         profileWeb.getMyUserGraphQlInfo<GetNicknameAndAvatarResponse>(
-            fields = setOf(GraphQLFieldDefinition.NickName,GraphQLFieldDefinition.Image,GraphQLFieldDefinition.Level,GraphQLFieldDefinition.Badge),
+            fields = setOf(
+                    GraphQLFieldDefinition.NickName,
+                    GraphQLFieldDefinition.Image,
+                    GraphQLFieldDefinition.Level,
+                    GraphQLFieldDefinition.Badge,
+                    GraphQLFieldDefinition.Bio
+            ),
             type = object : TypeToken<GetNicknameAndAvatarResponse>(){}.type
 
         ).logResponse(TAG)
         profileWeb.mutationMyUserGraphQlInfo<GetNicknameAndAvatarResponse>(
             type = object : TypeToken<GetNicknameAndAvatarResponse>(){}.type,
             variable = MutationData.Builder(
-                nickname = "泰瑞瑞瑞瑞",
-                image = "https://storage.googleapis.com/cmoney-image/1378ceeb-2f10-4ef5-8d38-cb63f8f97422"
+                nickname = "Tester_X",
+                image = "https://storage.googleapis.com/cmoney-image/1378ceeb-2f10-4ef5-8d38-cb63f8f97422",
+                bio = "我的自我介紹"
             ).build()
+        ).logResponse(TAG)
+        profileWeb.getMyUserGraphQlInfo<GetNicknameAndAvatarResponse>(
+                fields = setOf(
+                        GraphQLFieldDefinition.NickName,
+                        GraphQLFieldDefinition.Image,
+                        GraphQLFieldDefinition.Level,
+                        GraphQLFieldDefinition.Badge
+                ),
+                type = object : TypeToken<GetNicknameAndAvatarResponse>(){}.type
+
         ).logResponse(TAG)
         profileWeb.getUserGraphQLInfo<GetNicknameAndAvatarResponse>(
             memberIds = listOf(
@@ -71,7 +91,8 @@ class ProfileServiceCase : ServiceCase {
             fields = setOf(
                 UserGraphQLInfo.NickName,
                 UserGraphQLInfo.ID,
-                UserGraphQLInfo.Image
+                UserGraphQLInfo.Image,
+                UserGraphQLInfo.Bio
             ),
             //注意，因為回傳會是List，記得在型別要先加上List -> TypeToken<List<T>>(){}.type
             type = object : TypeToken<List<GetNicknameAndAvatarResponse>>(){}.type