+++ backend2/src/main/java/com/cmoney/backend2/chipk/service/api/internationalkchart/TickInfoSet.kt
@@ -1,0 +1,9 @@
+package com.cmoney.backend2.chipk.service.api.internationalkchart
+
+
+import com.google.gson.annotations.SerializedName
+
+data class TickInfoSet(
+    @SerializedName("TickInfoSet")
+    val tickInfoSet: List<TickInfo?>?
+)