+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/article/create/variable/commoditytag/CommodityTag.kt
@@ -3,10 +3,17 @@ package com.cmoney.backend2.forumocean.service.api.article.create.variable.commo
 import com.google.gson.annotations.SerializedName
 
 data class CommodityTag(
-    @SerializedName("commodityKey")
+    @SerializedName("key")
     val commodityKey : String,
-    @SerializedName("bullOrBear")
+    @Transient
     val bullOrBear : BullOrBear,
     @SerializedName("type")
     val type : StockTypeInfo
-)
+){
+
+    /**
+     * 為了用於解析欄位用 因為需要序列化成Int型態 所以無法直接在enum上使用SerializedName
+     */
+    @SerializedName("bullOrBear")
+    private val bullOrBearField : Int = bullOrBear.value
+}