+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/variable/response/articleresponse/contentoption/ArticleStateInfo.kt
@@ -1,0 +1,16 @@
+package com.cmoney.backend2.forumocean.service.api.variable.response.articleresponse.contentoption
+
+import com.cmoney.backend2.forumocean.service.api.variable.response.articleresponse.articlestate.ArticleState
+
+/**
+ * 請加入以下欄位
+ * @SerializedName("articleState")
+ * val articleState : ArticleState?
+ *
+ */
+interface ArticleStateInfo {
+    /**
+     * 文章狀態
+     */
+    val articleState : ArticleState?
+}