+++ backend2/src/main/java/com/cmoney/backend2/forumocean/service/api/variable/response/commentresponse/CommentContent.kt
@@ -3,6 +3,20 @@ package com.cmoney.backend2.forumocean.service.api.variable.response.commentresp
 import com.cmoney.backend2.forumocean.service.api.variable.response.mediatype.MediaTypeInfo
 import com.google.gson.annotations.SerializedName
 
+/**
+ * 回文內容
+ *
+ * @property creatorId 發文者
+ * @property text 內容
+ * @property multiMedia 多媒體資訊
+ * @property position 社團位置(官方回文才會帶)
+ * @property commentState 回文屬性
+ * @property reactionState 會員反應(自己的反應)
+ * @property reaction 反應類別 對應 反應總數
+ * @property createTime 發回文時間
+ * @property modifyTime 修改回文時間
+ * @property report 是否有檢舉
+ */
 data class CommentContent(
     @SerializedName("creatorId")
     val creatorId : Long?,
@@ -21,5 +35,7 @@ data class CommentContent(
     @SerializedName("createTime")
     val createTime : Long?,
     @SerializedName("modifyTime")
-    val modifyTime : Long?
+    val modifyTime : Long?,
+    @SerializedName("report")
+    val report: Any?
 )